+++ windivert-sys/src/bindings/mod.rs
@@ -15,10 +15,6 @@ pub use error::*;
 mod newtypes;
 pub use newtypes::*;
 
-use windows::Win32::{
-    Foundation::{BOOL, HANDLE},
-    System::IO::OVERLAPPED,
-};
 /// Default value for queue length parameter.
 pub const WINDIVERT_PARAM_QUEUE_LENGTH_DEFAULT: u64 = 4096;
 /// Minimum valid value for queue length parameter.
@@ -53,61 +49,61 @@ extern "C" {
         layer: WinDivertLayer,
         priority: i16,
         flags: WinDivertFlags,
-    ) -> HANDLE;
+    ) -> *mut c_void;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_recv)
     pub fn WinDivertRecv(
-        handle: HANDLE,
+        handle: *mut c_void,
         pPacket: *mut ::std::os::raw::c_void,
         packetLen: u32,
         pRecvLen: *mut u32,
         pAddr: *mut address::WINDIVERT_ADDRESS,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_recv_ex)
     pub fn WinDivertRecvEx(
-        handle: HANDLE,
+        handle: *mut c_void,
         pPacket: *mut ::std::os::raw::c_void,
         packetLen: u32,
         pRecvLen: *mut u32,
         flags: u64,
         pAddr: *mut address::WINDIVERT_ADDRESS,
         pAddrLen: *mut u32,
-        lpOverlapped: *mut OVERLAPPED,
-    ) -> BOOL;
+        lpOverlapped: *mut c_void,
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_send)
     pub fn WinDivertSend(
-        handle: HANDLE,
+        handle: *mut c_void,
         pPacket: *const ::std::os::raw::c_void,
         packetLen: u32,
         pSendLen: *mut u32,
         pAddr: *const address::WINDIVERT_ADDRESS,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_send_ex)
     pub fn WinDivertSendEx(
-        handle: HANDLE,
+        handle: *mut c_void,
         pPacket: *const ::std::os::raw::c_void,
         packetLen: u32,
         pSendLen: *mut u32,
         flags: u64,
         pAddr: *const address::WINDIVERT_ADDRESS,
         addrLen: u32,
-        lpOverlapped: *mut OVERLAPPED,
-    ) -> BOOL;
+        lpOverlapped: *mut c_void,
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_shutdown)
-    pub fn WinDivertShutdown(handle: HANDLE, how: WinDivertShutdownMode) -> BOOL;
+    pub fn WinDivertShutdown(handle: *mut c_void, how: WinDivertShutdownMode) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_close)
-    pub fn WinDivertClose(handle: HANDLE) -> BOOL;
+    pub fn WinDivertClose(handle: *mut c_void) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_set_param)
-    pub fn WinDivertSetParam(handle: HANDLE, param: WinDivertParam, value: u64) -> BOOL;
+    pub fn WinDivertSetParam(handle: *mut c_void, param: WinDivertParam, value: u64) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_get_param)
-    pub fn WinDivertGetParam(handle: HANDLE, param: WinDivertParam, pValue: *mut u64) -> BOOL;
+    pub fn WinDivertGetParam(handle: *mut c_void, param: WinDivertParam, pValue: *mut u64) -> i32;
 }
 
 extern "C" {
@@ -126,7 +122,7 @@ extern "C" {
         pDataLen: *mut u32,
         ppNext: *mut c_void,
         pNextLen: *mut u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_hash_packet)
     pub fn WinDivertHelperHashPacket(
@@ -139,27 +135,27 @@ extern "C" {
     pub fn WinDivertHelperParseIPv4Address(
         addrStr: *const ::std::os::raw::c_char,
         pAddr: *mut u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_parse_ipv6_address)
     pub fn WinDivertHelperParseIPv6Address(
         addrStr: *const ::std::os::raw::c_char,
         pAddr: *mut u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_format_ipv4_address)
     pub fn WinDivertHelperFormatIPv4Address(
         addr: u32,
         buffer: *mut ::std::os::raw::c_char,
         bufLen: u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_format_ipv6_address)
     pub fn WinDivertHelperFormatIPv6Address(
         pAddr: *const u32,
         buffer: *mut ::std::os::raw::c_char,
         bufLen: u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_calc_checksums)
     pub fn WinDivertHelperCalcChecksums(
@@ -167,13 +163,13 @@ extern "C" {
         packetLen: u32,
         pAddr: *mut address::WINDIVERT_ADDRESS,
         flags: ChecksumFlags,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_dec_ttl)
     pub fn WinDivertHelperDecrementTTL(
         pPacket: *mut ::std::os::raw::c_void,
         packetLen: u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_compile_filter)
     pub fn WinDivertHelperCompileFilter(
@@ -183,7 +179,7 @@ extern "C" {
         objLen: u32,
         errorStr: *mut *const ::std::os::raw::c_char,
         errorPos: *mut u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_eval_filter)
     pub fn WinDivertHelperEvalFilter(
@@ -191,7 +187,7 @@ extern "C" {
         pPacket: *const ::std::os::raw::c_void,
         packetLen: u32,
         pAddr: *const address::WINDIVERT_ADDRESS,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_format_filter)
     pub fn WinDivertHelperFormatFilter(
@@ -199,7 +195,7 @@ extern "C" {
         layer: WinDivertLayer,
         buffer: *mut ::std::os::raw::c_char,
         bufLen: u32,
-    ) -> BOOL;
+    ) -> i32;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_ntoh)
     pub fn WinDivertHelperNtohs(x: u16) -> u16;