+++ windivert/src/packet/full.rs
@@ -1,4 +1,5 @@
 use windivert_sys::{ChecksumFlags, WinDivertHelperCalcChecksums};
+use windows::Win32::Foundation::BOOL;
 
 use crate::{address::WinDivertAddress, layer, prelude::WinDivertError};
 
@@ -33,12 +34,12 @@ impl<'a> WinDivertPacket<'a, layer::NetworkLayer> {
     pub fn recalculate_checksums(&mut self, flags: ChecksumFlags) -> Result<(), WinDivertError> {
         if let Cow::Owned(ref mut data) = self.data.borrow_mut() {
             let res = unsafe {
-                WinDivertHelperCalcChecksums(
+                BOOL(WinDivertHelperCalcChecksums(
                     data.as_mut_ptr() as *mut c_void,
                     data.len() as u32,
                     self.address.as_mut(),
                     flags,
-                )
+                ))
             };
             if !res.as_bool() {
                 return Err(WinDivertError::from(windows::core::Error::from_win32()));
@@ -74,12 +75,12 @@ impl<'a> WinDivertPacket<'a, layer::ForwardLayer> {
     pub fn recalculate_checksums(&mut self, flags: ChecksumFlags) -> Result<(), WinDivertError> {
         if let Cow::Owned(ref mut data) = self.data.borrow_mut() {
             let res = unsafe {
-                WinDivertHelperCalcChecksums(
+                BOOL(WinDivertHelperCalcChecksums(
                     data.as_mut_ptr() as *mut c_void,
                     data.len() as u32,
                     self.address.as_mut(),
                     flags,
-                )
+                ))
             };
             if !res.as_bool() {
                 return Err(WinDivertError::from(windows::core::Error::from_win32()));